requirements.txt
# Bibliotecas de Simulação Dinâmica e Modelagem (Biologia de Sistemas)
copasi-basico>=0.85
tellurium>=2.2.11.1
pysb>=1.15.2

# Pacotes Científicos e de Dados (Base para Simulação e Plotagem)
pandas>=2.2.2
numpy>=1.26.4
scipy>=1.13.1
matplotlib>=3.9.0

# Ferramentas de Desenvolvimento e Utilidades
requests>=2.32.3
httpx>=0.27.2
PyYAML>=6.0.2
python-dotenv>=1.0.1
tenacity>=8.5.0
tqdm>=4.66.4
pytest>=8.3.2